python/examples/requirements.txt
PyYAML >= 5.4.1
./guardai_api
